refactor: modal 훅 분리
